Saas-App-backend/requirements.txt
flask
flask-cors
rembg[cpu]
pillow
spacy
opencv-python-headless
numpy
onnxruntime
gradio